--- airbyte-integrations/connectors/destination-snowflake/src/main/java/io/airbyte/integrations/destination/snowflake/SnowflakeInsertDestination.java
@@ -1,49 +1,0 @@
-/*
- * Copyright (c) 2022 Airbyte, Inc., all rights reserved.
- */
-
-package io.airbyte.integrations.destination.snowflake;
-
-import com.fasterxml.jackson.databind.JsonNode;
-import io.airbyte.commons.json.Jsons;
-import io.airbyte.db.jdbc.JdbcDatabase;
-import io.airbyte.integrations.base.Destination;
-import io.airbyte.integrations.destination.jdbc.AbstractJdbcDestination;
-import io.airbyte.integrations.util.OssCloudEnvVarConsts;
-import java.util.Collections;
-import java.util.Map;
-import javax.sql.DataSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class SnowflakeInsertDestination extends AbstractJdbcDestination implements Destination {
-
-  private static final Logger LOGGER = LoggerFactory.getLogger(SnowflakeDestination.class);
-
-  public SnowflakeInsertDestination() {
-    // the driver class is a no op because we override getDatabase.
-    super("", new SnowflakeSQLNameTransformer(), new SnowflakeSqlOperations());
-  }
-
-  @Override
-  protected DataSource getDataSource(final JsonNode config) {
-    return SnowflakeDatabase.createDataSource(config, OssCloudEnvVarConsts.AIRBYTE_OSS);
-  }
-
-  @Override
-  protected JdbcDatabase getDatabase(final DataSource dataSource) {
-    return SnowflakeDatabase.getDatabase(dataSource);
-  }
-
-  @Override
-  protected Map<String, String> getDefaultConnectionProperties(final JsonNode config) {
-    return Collections.emptyMap();
-  }
-
-  // this is a no op since we override getDatabase.
-  @Override
-  public JsonNode toJdbcConfig(final JsonNode config) {
-    return Jsons.emptyObject();
-  }
-
-}
